+++ hook.php
@@ -377,7 +377,7 @@ function plugin_fields_addWhere($link, $nott, $itemtype, $ID, $val, $searchtype)
                 $link .= ' NOT ';
             }
 
-            return $link . 'CAST(' . $DB->quoteName($table . '_' . $field) . '.' . $DB->quoteName($field) . ' AS DECIMAL(10,7))' . $operator . ' ' . $DB->quoteValue($val) ;
+            return $link . ' CAST(' . $DB->quoteName($table . '_' . $field) . '.' . $DB->quoteName($field) . ' AS DECIMAL(10,7))' . $operator . ' ' . $DB->quoteValue($val);
         } else {
             // if 'number' field with name is found with <= or >= or < or > search
             // update WHERE clause with the correct operator