+++ inc/field.class.php
@@ -355,9 +355,10 @@ class PluginFieldsField extends CommonDBChild
         $question = new Question();
         $found = $question->find([
             'type' => PluginFieldsQuestionType::class,
-            $this->fields['id'] => new QueryExpression(sprintf(
-                "JSON_VALUE(%s, '$.field_id')",
+            new QueryExpression(sprintf(
+                "JSON_VALUE(%s, '$.field_id') = %s",
                 DBmysql::quoteName('extra_data'),
+                DBmysql::quoteValue((int) $this->fields['id']),
             )),
         ]);
         if (!empty($found)) {
@@ -973,7 +974,7 @@ class PluginFieldsField extends CommonDBChild
         $itemtypes = PluginFieldsContainer::getUsedItemtypes($type, true);
 
         //if no dom containers defined for this itemtype, do nothing (in_array case insensitive)
-        if (!in_array(strtolower((string) $item::getType()), array_map(strtolower(...), $itemtypes))) {
+        if (!in_array(strtolower((string) $item::getType()), array_map(strtolower(...), $itemtypes), true)) {
             return;
         }